win32: HRESULT must be uintptr-sized for syscall.NewCallback (fixes Windows startup)
Running on Windows panicked in win32.init() —
'compileCallback: expected function with one uintptr-sized result' — because the
COM drag-drop vtable callbacks (IDropTarget/IDataObject/IDropSource) return
HRESULT, which was uint32 (4 bytes); syscall.NewCallback (strict on Go 1.25)
requires a uintptr-sized (8-byte) result. HRESULT is only ever a return value or
local (never a struct field) and the vtable slots are already uintptr, so
widening HRESULT to uintptr is safe and fixes every callback. Verified with
GOOS=windows go build ./win32/.

+++ win32/typedef.go
@@ -205,7 +205,7 @@ type (
 	HMODULE         HANDLE
 	HMONITOR        HANDLE
 	HPEN            HANDLE
-	HRESULT         uint32
+	HRESULT         uintptr
 	HRGN            HANDLE
 	HRSRC           HANDLE
 	HTHUMBNAIL      HANDLE
